requirements.txt
aiogram
psycopg2-binary
redis
python-dotenv
requests
